% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curr.R
\name{curr}
\alias{curr}
\title{Extract currency (or currencies) from a cell.}
\usage{
curr(L_i, L_max_i, n_i, h_i, p_i, f_i, d_i, v_i, beta_i, H_i, c_i, c_f,
  whatCurr_i, mu, l, e, NumFls, sumAll = TRUE, forageType = "random",
  alpha = 5e-05)
}
\arguments{
\item{L_i}{Load size (\eqn{\muL})}

\item{L_max_i}{Maximum load size (\eqn{\muL})}

\item{n_i}{Number of foragers in the cell}

\item{h_i}{Handling time per flower (s)}

\item{p_i}{Licking speed for nectar (\eqn{\muL/s})}

\item{f_i}{Flight time between flowers (s)}

\item{d_i}{Distance from hive (m)}

\item{v_i}{Unloaded flight speed from hive (m/s)}

\item{beta_i}{Decrease in flight speed with load (m/s\eqn{\muL})}

\item{H_i}{Time spent inside hive (s)}

\item{c_i}{Cost of non-flying behaviour (J/s)}

\item{c_f}{Cost of flight (J/s)}

\item{whatCurr_i}{Currency to use. Must be either "rat" (net rate) or "eff" (efficiency)}

\item{mu}{Per-flower nectar production (\eqn{\muL}/s)}

\item{l}{Maximum standing crop per flower (\eqn{\muL})}

\item{e}{Energetic value of nectar (J/\eqn{\muL})}

\item{NumFls}{Number of flowers per patch}

\item{sumAll}{Should currencies for the cell be summed? (useful for optimization of L) If not, returns a vector of currencies.}

\item{forageType}{Foraging style (see details below).}

\item{alpha}{Increase in flight cost with increase in load (5e-5 by default)}
}
\value{
Summed currency, or vector of currencies
}
\description{
Function to calculate currency(-ies) in a given cell. Called by \code{optimLoadCurr}
}
\details{
\code{forageType} can be one of the following:
\itemize{
\item \code{random}: random foraging (Possingham 1988)
\item \code{nn}: nearest-neighbour foraging
\item \code{omniscient}: flowers are shared optimally between foragers. Non-saturating nectar production.
\item \code{random_nonsat}: random foraging with non-saturating nectar production (ceiling at \code{l}).
}
}
\examples{
#Parameters
params=list(L_i=59.5,
           L_max_i=59.5,
           n_i=10,
           h_i=1.5,
           p_i=1,
           f_i=0.86,
           d_i=100,
           v_i=7.8,
           beta_i=0.102,
           H_i=100,
           c_f=0.05,
           c_i=0.0042,
           mu=0.3/3600,
           l=1,
           e=14.35,
           NumFls=520*(10^2))

temp1 <- with(params,
    curr(L_i,L_max_i,n_i,h_i,p_i,f_i,d_i,v_i,beta_i,H_i,
             c_i,c_f,mu,l,e,NumFls,whatCurr_i='eff',sumAll=FALSE,
             forageType='omniscient'))

temp1[['eff']] #Efficiency within cell
temp1[['S']] #Reduction in per-flower value
}
