% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCurr.R
\name{getCurr}
\alias{getCurr}
\title{Get currency (or currencies)}
\usage{
getCurr(
  whatCurr,
  L,
  L_max,
  e,
  d,
  v,
  h,
  f,
  l,
  p_i,
  c_i,
  c_f,
  H,
  alphaVal,
  betaVal,
  S
)
}
\arguments{
\item{whatCurr}{Currency to use. Must be either \code{"rat"} (net rate) or \code{"eff"} (efficiency)}

\item{L}{Load size (\eqn{\muL})}

\item{L_max}{Maximum load size (\eqn{\muL})}

\item{e}{Energetic value of nectar (J/\eqn{\muL})}

\item{d}{Distance from hive (m)}

\item{v}{Unloaded flight speed from hive (m/s)}

\item{h}{Handling time per flower (s)}

\item{f}{Flight time between flowers (s)}

\item{l}{Maximum standing crop per flower (\eqn{\muL})}

\item{p_i}{Licking speed for nectar (\eqn{\muL/s})}

\item{c_i}{Cost of non-flying behaviour (J/s)}

\item{c_f}{Cost of flight (J/s)}

\item{H}{Time spent inside hive (s)}

\item{alphaVal}{Increase in metabolic rate with load (J/(s*\eqn{\muL}))}

\item{betaVal}{Decrease in flight speed with load (m/s\eqn{\muL})}

\item{S}{Competition term (0-1)}
}
\value{
Currency, or vector of currencies
}
\description{
Function to calculate currency (or currencies) in a given cell. Called by \code{curr} to return nest-specific currency.
}
\examples{
#Parameters
params=list(L=59.5,
L_max=59.5,
e=14,
d=100,
v=7.8,
h=1.5,
f=0.86,
l=1,
p_i=1,
c_i=0.0042,
c_f=0.05,
H=100,
alphaVal=0.013,
betaVal=0.102/59.5)

with(params,getCurr(whatCurr='rat',L,L_max,e,d,v,h,f,
           l,p_i,c_i,c_f,H,betaVal,S=1))
with(params,getCurr(whatCurr='rat',L,L_max,e,d,v,h,f,
           l,p_i,c_i,c_f,H,betaVal,S=0.5))

}
