% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeBest.R
\name{makeBest}
\alias{makeBest}
\title{Create scenario with +TRANSFER more foragers}
\usage{
makeBest(scenario, whichNest = NA, parallel = F, cluster = NA)
}
\arguments{
\item{scenario}{Nest structure and world structure (scenario)}

\item{whichNest}{Which nest should have TRANSFER foragers taken away?}

\item{parallel}{Should computation be done in parallel?}

\item{cluster}{If parallel, which cluster should be used}
}
\value{
List of nests and world structure (scenario)
}
\description{
Create and optimize currency in a world where -\code{transfer} foragers has been taken from nest(s) \code{whichnest}. Called within \code{forageMod}
}
\examples{
makeBest(base,whichNest=1,parallel=T,cluster=cluster)
}
