% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeWorst.R
\name{makeWorst}
\alias{makeWorst}
\title{World with -TRANSFER less foragers}
\usage{
makeWorst(nests, world, whichNest = NA, parallel = F, cluster = NA)
}
\arguments{
\item{nests}{Nest structure}

\item{world}{World structure}

\item{whichNest}{Which nest should have TRANSFER foragers taken away?}

\item{parallel}{Should computation be done in parallel?}

\item{cluster}{If parallel, which cluster should be used}
}
\value{
List of nests and world structure
}
\description{
World with -TRANSFER less foragers
}
\examples{
#Parameters
temp<-makeWorst(nests,world,1,T,cluster)
temp$nests #Nests
temp$world #World
}
