% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moveForagers.R
\name{moveForagers}
\alias{moveForagers}
\title{Move foragers from one cell to another, and re-calculate L, S, and curr}
\usage{
moveForagers(scenarioSet, moves)
}
\arguments{
\item{scenarioSet}{Scenario set list, containing best, base, and worst
scenarios (worst is NA for solitary foragers)}

\item{moves}{Move list from whichMoves, telling which cells to move from and
to.}
}
\value{
Updated scenario set list
}
\description{
\code{moveForagers} moves foragers from one cell to another, and
 re-calculates L, S, and curr.
}
\examples{
#Create scenario
nests<-list(nest1=list(xloc=1,yloc=1,n=matrix(c(5,10,15),1),
                    whatCurr="eff",sol=T,
                    eps=0,L_max=59.5,v=7.8,
                    beta=0.102/59.5,p_i=1,
                    h=matrix(rep(1.5,3),1),
                    c_f=0.05,c_i=0.0042,
                    H=100,d=matrix(c(50,100,150),1),
                    L=matrix(rep(59.5,3),1),
                    curr=matrix(c(5,10,15),1),
                    steps=5,stepNum=1))

world<-list(mu=matrix(rep(8.33e-05,3),1),
         flDens=matrix(rep(520,3),1),
         e=matrix(rep(14.3,3),1),
         l=matrix(rep(1,3),1),
         f=matrix(rep(0.86,3),1),
         cellSize=10,patchLev=FALSE,
         S=matrix(rep(1,3),1),
         forageType='random')

baseScen=list(nests=nests,world=world) #Baseline scenario
}
