% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimLoadCurr.R
\name{optimLoadCurr}
\alias{optimLoadCurr}
\title{Optimal currency and load}
\usage{
optimLoadCurr(u, scenario)
}
\arguments{
\item{u}{Address of cells in Scenario to calculate}

\item{scenario}{Scenario to use}
}
\value{
List of optimal currency and load values for each cell in \code{u}
}
\description{
Function to calculate optimal currency and load in cell u. Returns NA values
if arguments to cells are NA. This is the "workhorse" function of CPForage
}
\examples{

}
