% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scalFun.R
\name{scalFun}
\alias{scalFun}
\title{scalFun}
\usage{
scalFun(mu = NULL, l = NULL, NumFls = NULL, L_i = NULL,
  L_max_i = NULL, n_i = NULL, h_i = NULL, p_i = NULL, f_i = NULL,
  d_i = NULL, v_i = NULL, beta_i = NULL, H_i = NULL,
  forageType = NULL)
}
\arguments{
\item{mu}{Per-flower nectar production (\eqn{\muL}/s)}

\item{l}{Maximum standing crop per flower (\eqn{\muL})}

\item{NumFls}{Number of flowers per patch}

\item{L_i}{Load size (\eqn{\muL})}

\item{L_max_i}{Maximum load size (\eqn{\muL})#'}

\item{n_i}{Number of foragers in the cell}

\item{h_i}{Handling time per flower (s)}

\item{p_i}{Licking speed for nectar (\eqn{\muL/s})}

\item{f_i}{Flight time between flowers (s)}

\item{d_i}{Distance from hive (m)}

\item{v_i}{Unloaded flight speed from hive (m/s)}

\item{beta_i}{Decrease in flight speed with load (m/s\eqn{\muL})}

\item{H_i}{Time spent inside hive (s)}

\item{forageType}{Type of foraging within patch (see details)}
}
\value{
S-value for patch
}
\description{
Function to find  S (scaling term), given Load, Forager number, Max intake rate, travel time, and nectar production
}
\details{
\code{forageType} can be one of the following:
\itemize{
\item \code{random}: random foraging (Possingham 1988)
\item \code{multiNN}: nearest-neighbour foraging
\item \code{omniscient}: flowers are shared optimally between foragers. Non-saturating nectar production.
\item \code{random_nonsat}: random foraging with non-saturating nectar production (ceiling at \code{l}).

}
}
\examples{
#Test
mu=0.3/3600
l=1
NumFls=520 #520 flowers
L_i=59.5
L_max_i=59.5
n_i=10
h_i=1.5
p_i=1
f_i=0.86
d_i=100
v_i=7.8
beta_i=0.102
H_i=100
forageType='omniscient'

scalFun(mu,l,NumFls,L_i,L_max_i,n_i,h_i,p_i,f_i,
                d_i,v_i,beta_i,H_i,forageType)
}
