% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeWorst.R
\name{makeWorst}
\alias{makeWorst}
\title{Create scenario with -TRANSFER less foragers}
\usage{
makeWorst(scenario, parallel = F, cluster = NA)
}
\arguments{
\item{scenario}{Nest structure and world structure (scenario)}

\item{parallel}{Should computation be done in parallel?}

\item{cluster}{If parallel, which cluster should be used}
}
\value{
List of nests and world structure (scenario)
}
\description{
Create and optimize currency in a world where -TRANSFER foragers has been taken from nest(s) WHICHNEST. Called within \code{forageMod}
}
\examples{
makeWorst(base,parallel=T,cluster=cluster)
}
