% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeWorst.R
\name{makeWorst}
\alias{makeWorst}
\title{World with -TRANSFER less foragers}
\usage{
makeWorst(scenario, whichNest = NA, parallel = F, cluster = NA)
}
\arguments{
\item{scenario}{Nest structure and world structure (scenario)}

\item{whichNest}{Which nest should have TRANSFER foragers taken away?}

\item{parallel}{Should computation be done in parallel?}

\item{cluster}{If parallel, which cluster should be used}
}
\value{
List of nests and world structure (scenario)
}
\description{
World with -TRANSFER less foragers
}
